Login height reduced | email should not be updated from profile

+++ lib/features/auth/domain/entities/user_detail.dart
@@ -61,7 +61,7 @@ class UserDetail extends Equatable {
   final String? mobile;
   final String? uuid;
   final String? avatar;
-  final int? balance;
+  final double? balance;
   final int? point;
   final String? dob;
   final String? gender;
